service/requirements.txt
dlib==19.20.0
face-recognition==1.3.0
Flask==1.1.2
tensorflow>=1.15.2
keras==2.3.1
imutils==0.5.3
numpy==1.18.2
opencv-python==4.2.0.*
matplotlib==3.2.1
argparse==1.1
scipy==1.4.1
scikit-learn==0.23.1
pillow==7.2.0
absl-py==0.4.0
astor==0.7.1
ffmpy==0.2.2
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
Markdown==2.6.11
munch==2.3.2
nvidia-ml-py3==7.352.0
olefile==0.46
pathlib==1.0.1
protobuf
pycparser==2.19
scandir
six==1.12.0
termcolor==1.1.0
tqdm==4.48.2
pandas==1.1.0
pretrainedmodels==0.7.4
